+++ src/test/java/br/com/project/checkskills/test/repositories/CargoRepositoryTest.java
@@ -12,7 +12,7 @@ import br.com.project.checkskills.entities.dadosbasicos.CargoEntity;
 import br.com.project.checkskills.entities.dadosbasicos.NivelEntity;
 import br.com.project.checkskills.repositories.dadosbasicos.ICargoRepository;
 import br.com.project.checkskills.repositories.dadosbasicos.INivelRepository;
-import br.com.project.checkskills.utils.AbstractDatabaseTest;
+import br.com.project.checkskills.test.utils.AbstractDatabaseTest;
 
 
 public class CargoRepositoryTest extends AbstractDatabaseTest {
