+++ src/test/java/br/com/project/checkskills/test/repositories/DepartamentoRepositoryTest.java
@@ -10,9 +10,9 @@ import org.junit.Test;
 import br.com.project.checkskills.entities.dadosbasicos.DepartamentoEntity;
 import br.com.project.checkskills.entities.dadosbasicos.OrganizacaoEntity;
 import br.com.project.checkskills.repositories.dadosbasicos.IDepartamentoRepository;
+import br.com.project.checkskills.test.utils.AbstractDatabaseTest;
 
-
-public class DepartamentoRepositoryTest  {
+public class DepartamentoRepositoryTest extends AbstractDatabaseTest {
 	
 	private static final Logger LOGGER = Logger.getLogger(DepartamentoRepositoryTest.class);
