+++ src/test/java/br/com/project/checkskills/test/utils/AbstractDatabaseTest.java
@@ -1,0 +1,10 @@
+package br.com.project.checkskills.test.utils;
+
+import org.junit.runner.RunWith;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@ContextConfiguration(locations = { "/applicationContext-test.xml" })
+public abstract class AbstractDatabaseTest {
+}
